% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ac.inGroups}
\alias{ac.inGroups}
\title{Check if ag/sr is in groups}
\usage{
ac.inGroups(ac, ...)
}
\arguments{
\item{...}{groups to check}
}
\value{
char
}
\description{
Check if ag/sr is in groups
}
