% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{acdb.applyFunction}
\alias{acdb.applyFunction}
\title{Apply a function over an acdb}
\usage{
acdb.applyFunction(acdb, fn, ..., unlist = F)
}
\arguments{
\item{acdb}{list}

\item{fn}{function}
}
\value{
list
}
\description{
Efficiently applies a function over acdb.
}
