% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_methods.R
\name{acdb.as.list}
\alias{acdb.as.list}
\title{Convert database entries to lists}
\usage{
acdb.as.list(db)
}
\description{
Antigen entries become lists of properties rather than environments \code{.parent} entries are removed.
}
