% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{acdb.extract}
\alias{acdb.extract}
\title{Match and extract a single antigen from attributes}
\usage{
acdb.extract(db, ..., inherit = TRUE, inherit.gene = "HA")
}
\arguments{
\item{db}{list: an agdb (see \code{?agdb})}

\item{...}{attributes to match}

\item{inherit.gene}{char: which gene to inherit along}

\item{inherit:}{bool: whether to inherit attributes from parents}

\item{simplify}{bool: whether to simplify list of length 1 to \code{list[[1]]}}
}
\value{
list
}
\description{
Find an antigen entry which has attributes specified in '...'. An error is thrown if there are multiple matches. Attribute inheritance behavior can be set with \code{inherit} and \code{inherit.gene} (see \code{?ag.attribute}).
}
