% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{acdb.idToAttr}
\alias{acdb.idToAttr}
\title{Get named attribute from acdb}
\usage{
acdb.idToAttr(
  ids,
  attribute,
  inherit = T,
  inherit.gene = "HA",
  acdb = get_agdb()
)
}
\arguments{
\item{ids}{char}

\item{attribute}{char}

\item{acdb}{}
}
\value{
char
}
\description{
Get named attribute from acdb
}
