% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{acdb.matchNames}
\alias{acdb.matchNames}
\title{Find name matches in an antigen or serum database}
\usage{
acdb.matchNames(
  strain_names,
  acdb,
  include_aliases = TRUE,
  multiple_match_error = TRUE
)
}
\arguments{
\item{strain_names}{names to be matched}

\item{include_aliases}{should aliases be searched?}

\item{multiple_match_error}{should a name matchingmultiple db entries throw an error?}
}
\value{
Returns a list of standardised names and extracted information
}
\description{
Find name matches in an antigen or serum database
}
