% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{acdb.matchNamesIndices}
\alias{acdb.matchNamesIndices}
\title{Function to find db entries with names matching passed strain names
Useful for finding antigen match when adding new tables to database}
\usage{
acdb.matchNamesIndices(
  strain_names,
  acdb,
  include_aliases = TRUE,
  multiple_match_error = TRUE
)
}
\description{
Function to find db entries with names matching passed strain names
Useful for finding antigen match when adding new tables to database
}
