% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{acdb.matchSubstitutions}
\alias{acdb.matchSubstitutions}
\title{Find db entries with same subsititutions as subs}
\usage{
acdb.matchSubstitutions(ags, subs, gene = "HA", no_extras = TRUE)
}
\description{
Find db entries with same subsititutions as subs
}
