% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{acdb.nameIDs}
\alias{acdb.nameIDs}
\title{Get long names from acdb}
\usage{
acdb.nameIDs(ids, acdb, name = "long")
}
\arguments{
\item{ids}{char}

\item{acdb}{}
}
\value{
char
}
\description{
Get long names from acdb
}
