% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{acdb.search}
\alias{acdb.search}
\title{Match and extract attributes}
\usage{
acdb.search(db, ..., inherit = TRUE, inherit.gene = "HA", simplify = TRUE)
}
\arguments{
\item{db}{list: an agdb (see \code{?agdb})}

\item{...}{attributes to match}

\item{inherit.gene}{char: which gene to inherit along}

\item{simplify}{bool: whether to simplify list of length 1 to \code{list[[1]]}}

\item{inherit:}{bool: whether to inherit attributes from parents}
}
\value{
list
}
\description{
Find entries which have attributes specified in '...'. This function is same as agdb.search, and is curried to form srdb.search.
}
\details{
Attribute inheritance behavior can be set with \code{inherit} and \code{inherit.gene} (see \code{?ag.attribute})
}
