% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_methods.R
\name{acdb.slice}
\alias{acdb.slice}
\title{Subsetting for databases}
\usage{
acdb.slice(db, val)
}
\description{
Returns field for all database entries. Like \verb{\%$\%}, except variable val is supported.
}
