% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{acids.applyFunction}
\alias{acids.applyFunction}
\title{Apply a function over an acdb subset, found by id}
\usage{
acids.applyFunction(ids, acdb, fn)
}
\arguments{
\item{acdb}{list}

\item{fn}{function}
}
\value{
list
}
\description{
Efficiently applies a function over acdb.
}
