% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_functions.R
\name{ag.attribute}
\alias{ag.attribute}
\title{Get gene attribute}
\usage{
ag.attribute(ag, attribute, inherit = TRUE, inherit.gene = "HA")
}
\arguments{
\item{ag}{environment: an antigen}

\item{attribute}{char: the name of the desired attiribute}

\item{inherit.gene}{char: which gene to inherit along}

\item{inherit:}{bool: whether to inherit attributes from parents}
}
\value{
attribute value
}
\description{
Find an attribute value for an antigen. By default this is set to traverse the inheritance list for 'HA' to find the attribute value but the gene can be set to 'NA' or 'backbone', or this behaviour turned off.
}
