% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.children}
\alias{ag.children}
\title{Get children}
\usage{
ag.children(ag, agdb, gene = "BB", parents = NULL)
}
\arguments{
\item{ag}{environment: an antigen}
}
\value{
list
}
\description{
Get children
}
