% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.clade}
\alias{ag.clade}
\title{Get antigen root clade}
\usage{
ag.clade(ag, how = "self")
}
\arguments{
\item{how}{= c('self', 'root', 'any')}
}
\value{
char
}
\description{
Get antigen root clade
}
