% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.descendents}
\alias{ag.descendents}
\title{Get all descendants}
\usage{
ag.descendents(ag, agdb)
}
\arguments{
\item{ag}{environment: an antigen}
}
\value{
list
}
\description{
Gets all descendants of an antigen (backbone-wise), flattening to list (compare to \code{?ag.descendantsTree})
}
