% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.descendentsTree}
\alias{ag.descendentsTree}
\title{Get all descendants}
\usage{
ag.descendentsTree(ag, agdb, gene = "BB")
}
\arguments{
\item{ag}{environment: an antigen}
}
\value{
list
}
\description{
Gets all descendants of an antigen (backbone-wise), preserving tree of relationships (compare to \code{?ag.descendants})
}
