% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.experiments}
\alias{ag.experiments}
\title{Get experiments that an antigen is in}
\usage{
ag.experiments(ag, expdb, include_passage_variants = FALSE, agdb)
}
\description{
Get experiments that an antigen is in
}
