% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.inheritance}
\alias{ag.inheritance}
\title{Get inheritance list for a gene.}
\usage{
ag.inheritance(ag, gene, inheritance = list())
}
\arguments{
\item{ag}{environment: an antigen}

\item{gene}{char: which gene inheritance list is required (must be 'HA', 'NA', or 'backbone)}
}
\value{
list
}
\description{
Get a list of parent antigens for a gene in an antigen. The list will follow the parent_id's for the specified component ('HA', 'NA', or 'backbone') until reaching the root.
Passed antigen is last element of list
}
