% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.long}
\alias{ag.long}
\title{Get long name for an antigen}
\usage{
ag.long(ag, agdb = get_agdb())
}
\arguments{
\item{ag}{environment: an antigen}

\item{agdb}{list: an agdb}
}
\value{
list
}
\description{
Get long name for an antigen
}
