% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_passage.R
\name{ag.passageSiblings}
\alias{ag.passageSiblings}
\title{Get all antigens which differ only by passage}
\usage{
ag.passageSiblings(ag, agdb)
}
\description{
Get all antigens which differ only by passage
}
