% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{ag}
\alias{ag}
\title{Antigen}
\description{
Antigen
}
\details{
An antigen entry (ag) is an environment containing some of:

\tabular{ll}{
id   \tab  Antigen id: 6 random capital letters and numbers \cr
strain \tab  Rarely used. \cr
long \tab Full strain name. \cr
aliases \tab Other names known to be used to refer to this antigen. \cr
wildtype \tab Boolean: whether the strain is wildtype. \cr
type \tab Influenza virus type (A, B...) \cr
subtype \tab Influenza virus subtype (H3N2, H1N1...) \cr
lineage \tab Victoria etc \cr
isolation \tab list containing isolation id, location, date, cell, continent \cr
genes \tab gene sequence for HA/NA \cr
parent_id \tab id of parent (by backbone) \cr
.parent \tab the database entry for the parent antigen (by backbone) \cr
alterations \tab alterations to parent (substitutions or gene transplants), organised by gene. Substitutions formatted as X123Y, transgenes identified by parent id. alterattion$.parent is the database entry for the gene origin when the alteration is a transplant.  \cr
passage \tab passaging history \cr
comments \tab Extra comments - eg which study the antigen comes from \cr
groups \tab Useful annotations. For example: wildtype, mutatant, gen 1 root, gen 2 mutant, 3C.3A etc.  \cr
meta \tab Some extra data about the antigen: cluster, mutant generation number, whether it is a reference strain \cr
}

Notably, the parent environment is always the empty environment; $.parent is the backbone-wise parent antigen entry; alteration$.parent is the origin antigen for a transplanted gene

It has classes c("acdatabase.ag", "acdatabase.entry", "environment")
}
