% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.sequence.allVariants}
\alias{ag.sequence.allVariants}
\title{Calculate gene sequence}
\usage{
ag.sequence.allVariants(
  ag,
  gene = "HA",
  agdb = acutils:::get_agdb(),
  inherit = T,
  full_search = F
)
}
\arguments{
\item{ag}{environment: an antigen}

\item{gene}{char: which gene to get sequence for}
}
\value{
list
}
\description{
Returns tibble of all sequences found for an antigen, and their relationship to the passed antigen
}
