% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.sequence.calculate}
\alias{ag.sequence.calculate}
\title{Calculate sequence for a gene from an inheritance (by finding a sequence and applying substitutions)}
\usage{
ag.sequence.calculate(inheritance, gene)
}
\description{
Calculate sequence for a gene from an inheritance (by finding a sequence and applying substitutions)
}
