% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.sequence.full}
\alias{ag.sequence.full}
\title{Calculate gene sequence}
\usage{
ag.sequence.full(
  ag,
  gene = "HA",
  agdb = acutils:::get_agdb(),
  inherit = T,
  passage = T
)
}
\arguments{
\item{ag}{environment: an antigen}

\item{gene}{char: which gene to get sequence for}
}
\value{
list
}
\description{
Calculate gene sequence
}
