% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.sequence}
\alias{ag.sequence}
\title{Calculate gene sequence}
\usage{
ag.sequence(ag, gene = "HA", agdb = acutils:::get_agdb(), inherit = T)
}
\arguments{
\item{ag}{environment: an antigen}

\item{gene}{char: which gene to get sequence for}
}
\value{
list
}
\description{
Calculate gene sequence
}
