% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.short}
\alias{ag.short}
\title{Construct short name for an antigen}
\usage{
ag.short(ag, agdb = get_agdb())
}
\arguments{
\item{ag}{environment: an antigen}

\item{agdb}{list: an agdb}
}
\value{
list
}
\description{
Construct short name for an antigen
}
