% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.substitutions}
\alias{ag.substitutions}
\title{Get gene substituions for antigen}
\usage{
ag.substitutions(ag, gene, substitutions = list(), inherit = TRUE)
}
\arguments{
\item{ag}{environment: an antigen}

\item{gene}{char}

\item{inherit}{bool: whether to check parents of gene for substitutions}
}
\value{
list
}
\description{
Gets all gene substitutions since the root in specified gene
}
