% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.ag}
\alias{agdb.ag}
\title{Produce an antigen entry for the database}
\usage{
agdb.ag(
  id,
  strain,
  long,
  aliases,
  wildtype,
  type,
  subtype,
  lineage,
  isolation,
  genes,
  parent_id,
  alterations,
  passage,
  comments,
  groups,
  meta,
  agdb,
  .parent = NULL
)
}
\arguments{
\item{id}{should be left NULL}

\item{agdb}{list}

\item{excluded_ids}{character (optional) a vector of ids which should not be assigned}
}
\value{
character
}
\description{
Produces an entry for antigen databases (see ?agdb) from the passed data.
}
