% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.alterations}
\alias{agdb.alterations}
\title{Constructs  an antigen alterations entry}
\usage{
agdb.alterations(gene, parent_id, substitutions)
}
\arguments{
\item{gene}{char: gene name (HA or NA)}

\item{parent_id}{char: the id of the origin virus for a transgene}

\item{substitutions}{char: vector containing substitutions of form 'X123Y'
#'}
}
\value{
list
}
\description{
Constructs an antigen alterations entry. All params are optional
}
