% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.append}
\alias{agdb.append}
\title{Append an antigen to an agdb}
\usage{
agdb.append(agdb, ag)
}
\arguments{
\item{agdb}{list}

\item{ag}{environment}
}
\value{
list
}
\description{
Append an antigen to an agdb
}
