% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.checkAG}
\alias{agdb.checkAG}
\title{Check antigen entry formatting}
\usage{
agdb.checkAG(ag, null_permitted = F)
}
\arguments{
\item{ag}{environment: a proposed antigen entry}

\item{null_permitted}{bool: whether NULL should return T or F}
}
\value{
bool
}
\description{
Checks that a proposed antigen entry follows the formatting rules. See ?ag for more details on antigen objects.
}
\details{
The following are currently checked:
\enumerate{
\item class is c("acdatabase.ag", "acdatabase.entry", "environment")
\item only permitted fields are entered
\item all names (strain, long. type, subtype, lineage) are characters
\item construct checkers are fulfilled for : id, parent_id, alterations, genes, isolation, passage, .parent.
}
}
