% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.checkagdb}
\alias{agdb.checkagdb}
\title{Checks Antigen Database formatting}
\usage{
agdb.checkagdb(agdb)
}
\arguments{
\item{agdb}{list}
}
\value{
bool
}
\description{
Checks that an antigen database follows the formatting rules. For information on database structure, see ?agdb.
}
\details{
The following rules are currently checked:
\enumerate{
\item All entries are antigens (see ?agddb.checkAG)
}
}
