% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_dbtest.R
\name{agdb.checkgenes}
\alias{agdb.checkgenes}
\title{Check an antigen genes entry}
\usage{
agdb.checkgenes(genes)
}
\arguments{
\item{genes}{char: A proposed genes entry}

\item{null_permitted}{bool: whether NULL should return T or F}
}
\value{
character
}
\description{
Checks if a proposed genes entry follows the formatting rules.
}
\details{
The following rules are currently checked:
\enumerate{
\item genes is a list
\item only permitted fields are present for each gene in genes
\item all genes are HA or NA
}
}
