% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.checkid}
\alias{agdb.checkid}
\title{Check an id}
\usage{
agdb.checkid(id, null_permitted = F)
}
\arguments{
\item{id}{char: A proposed id}

\item{null_permitted}{bool: whether NULL should return T or F}
}
\value{
character
}
\description{
Checks if a proposed id follows the formatting rules.
}
\details{
The following rules are currently checked:
\enumerate{
\item id is be character type
\item id has length 6
\item all characters are capital letters or numerals 0:9
}
}
