% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.checkisolation}
\alias{agdb.checkisolation}
\title{Check an antigen isolation entry}
\usage{
agdb.checkisolation(isolation)
}
\arguments{
\item{isolation}{char: A proposed isolation entry}

\item{null_permitted}{bool: whether NULL should return T or F}
}
\value{
character
}
\description{
Checks if a proposed isolation entry follows the formatting rules.
}
\details{
The following rules are currently checked:
\enumerate{
\item isolation is a list
\item only permitted fields are present
}
}
