% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.checkpassage}
\alias{agdb.checkpassage}
\title{Check a passage entry}
\usage{
agdb.checkpassage(passage)
}
\arguments{
\item{passage}{char: A proposed passage entry}
}
\value{
character
}
\description{
Checks if a proposed passage entry follows the correct formatting
}
\details{
The following rules are currently checked:
\enumerate{
\item passage is of list type
\item only permitted names are present
}
}
