% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{agdb.experiments}
\alias{agdb.experiments}
\title{Get experiments that antigens are in}
\usage{
agdb.experiments(ags, expdb, include_passage_variants, agdb)
}
\description{
Get experiments that antigens are in
}
