% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.genes}
\alias{agdb.genes}
\title{Constructs  an antigen gene entry}
\usage{
agdb.genes(gene, sequence, clade, cluster)
}
\arguments{
\item{gene}{char: gene name (HA or NA)}

\item{sequence}{char}

\item{clade}{char}

\item{cluster}{char}
}
\value{
list
}
\description{
Constructs an antigen gene entry. All params are optional
}
