% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{agdb.getIDs}
\alias{agdb.getIDs}
\title{Subset antigen database based on IDs}
\usage{
agdb.getIDs(ids, agdb)
}
\arguments{
\item{ids}{}
}
\value{
list
}
\description{
Throws error if some IDs not found
}
