% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{agdb.hasSubstitutions}
\alias{agdb.hasSubstitutions}
\title{Are there subsitutions in the inheritance of HA?}
\usage{
agdb.hasSubstitutions(agdb, inherit = T)
}
\arguments{
\item{agdb}{}
}
\value{

}
\description{
Are there subsitutions in the inheritance of HA?
}
