% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{agdb.homologousSera}
\alias{agdb.homologousSera}
\title{Get homologous sera}
\usage{
agdb.homologousSera(agdb, srdb)
}
\arguments{
\item{srdb}{}
}
\value{
list
}
\description{
Gets homologous sera for all antigens in an agdb by matching ids.
}
