% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.id}
\alias{agdb.id}
\title{Generate an id}
\usage{
agdb.id(db, id = NULL, excluded_ids = NULL)
}
\arguments{
\item{id}{should be left NULL}

\item{excluded_ids}{char (optional): a vector of ids which should not be assigned}

\item{agdb}{list}
}
\value{
character
}
\description{
Generates a random id (6 capital letters and numbers) which is not seen in the provided agdb or excluded_ids
}
