% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.isolation}
\alias{agdb.isolation}
\title{Constructs  an antigen isolation entry}
\usage{
agdb.isolation(id, date, cell, location, continent)
}
\arguments{
\item{id}{char: the isolation id}

\item{date}{char: the year of the isolation}

\item{cell}{char: cell or eg passaging?}

\item{location}{char#}

\item{continent}{char}
}
\value{
list
}
\description{
Constructs an antigen isolation entry. All params are optional
}
