% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.merge}
\alias{agdb.merge}
\title{Merge antigens in agdb}
\usage{
agdb.merge(agdb, ag_id1, ag_id2)
}
\arguments{
\item{agdb}{list}

\item{ag_id_keep}{char}

\item{ag_id_remove}{char}
}
\value{
list
}
\description{
Merge antigens in agdb
}
