% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.new}
\alias{agdb.new}
\title{Create antigen database}
\usage{
agdb.new(db = list())
}
\arguments{
\item{db}{list}
}
\value{
environment
}
\description{
Creates an antigen database from a list of antigen entries. For information on database structure, see ?agdb.
}
\examples{
# load database
package_path = system.file(package = "acutilsLite")
db_path = paste0(package_path, "/tests/testthat/testdata/agdb_h3_small.json")
print(db_path)
library(jsonlite)
db_list = read_json(db_path)
db <- agdb.new(db_list)
}
