% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb.passage}
\alias{agdb.passage}
\title{Constructs  an antigen passage entry}
\usage{
agdb.passage(history, egg, cell, details, comments)
}
\arguments{
\item{history}{char: vecotr of known passage history}

\item{egg}{char: not used}

\item{cell}{char: cell or egg passaging}
}
\value{
list
}
\description{
Constructs an antigen passage entry. All params are optional
}
