% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{agdb}
\alias{agdb}
\title{Antigen database}
\description{
Antigen database
}
\details{
The antigen database is ultimately stored as a json file with fields:
\enumerate{
\item id
\item strain
\item long
\item aliases
\item wildtype
\item type
\item subtype
\item lineage
\item isolation
\item genes
\item parent_id
\item alterations
\item passage
\item comments
\item groups
\item meta
}

Not all fields are filled for all antigens.

In R, this is converted to a list of environments (see ?ag) with the above fields, plus:
\enumerate{
\item A ag$.parent entry, which is the database entry for the parent antigen (defined by the backbone).
\item ag$alteration$.parent entries for each gene that is not inherited from ag$.parent (for recombinant viruses).
}
}
