% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{agdb.sequences}
\alias{agdb.sequences}
\title{Get sequences from agdb}
\usage{
agdb.sequences(agdb, gene = "HA", inherit = T)
}
\arguments{
\item{agdb}{}
}
\value{

}
\description{
Get sequences from agdb
}
