% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{agdb.sequencesFromMap}
\alias{agdb.sequencesFromMap}
\title{Get sequences from agdb}
\usage{
agdb.sequencesFromMap(
  ags,
  map,
  gene = "HA",
  inherit = T,
  fn = ag.sequence,
  agdb = get_agdb()
)
}
\arguments{
\item{agdb}{}
}
\value{

}
\description{
Get sequences from agdb
}
