% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{agdb.short}
\alias{agdb.short}
\title{Generate short name from antigen entries}
\usage{
agdb.short(agdb)
}
\arguments{
\item{ags}{}
}
\value{
Returns a vector of short names
}
\description{
Generate short name from antigen entries
}
