% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{agdb.substitutions}
\alias{agdb.substitutions}
\title{Get substitutions from agdb}
\usage{
agdb.substitutions(agdb, inherit = T)
}
\arguments{
\item{agdb}{}
}
\value{

}
\description{
Get substitutions from agdb
}
