% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{agid.cluster}
\alias{agid.cluster}
\title{Get antigen cluster}
\usage{
agid.cluster(identifier, agdb, from = "id")
}
\arguments{
\item{identifier}{char: an id, long, etc}

\item{from:}{char: which attribute to search with (long, id, etc)}
}
\value{
char
}
\description{
Get an antigen's cluster from id / long
}
