% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_methods.R
\name{as.list.acdatabase.ag}
\alias{as.list.acdatabase.ag}
\title{Convert database entry to list}
\usage{
\method{as.list}{acdatabase.ag}(record, ...)
}
\description{
Antigen entries become lists of properties rather than environments. \code{.parent} entries are removed.
}
