% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{check_condition}
\alias{check_condition}
\title{Check if a condition is fulfilled}
\usage{
check_condition(checker_fn, ...)
}
\arguments{
\item{checker_fn}{function: a boolean valued function checking if a condition is fulfilled by the arguments ...}

\item{...}{arguments to checker_fn}
}
\value{
bool
}
\description{
Checks if checker_fn returns TRUE when passed arguments '...'. Otherwise throws error.
}
