% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{collate}
\alias{collate}
\title{List to vector}
\usage{
collate(x, null = NA)
}
\arguments{
\item{x}{list}
}
\value{
vector
}
\description{
Converts list of length 1 vectors to vector. Throws error if an element has length > 1
}
