% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_merging.R
\name{combineTiters}
\alias{combineTiters}
\title{Combine titers}
\usage{
combineTiters(titers)
}
\arguments{
\item{titers}{char or int}
}
\value{
char: combined titers
}
\description{
Combine vector of titers to "240/480/240"-like
}
