% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_parsing.R
\name{exp.longTibble}
\alias{exp.longTibble}
\title{Parse an experiment to tibble}
\usage{
\method{exp}{longTibble}(exp)
}
\arguments{
\item{exp}{}
}
\value{
tibble
}
\description{
Parse an experiment record into a long tibble containing all results.
}
