% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_merging.R
\name{exp.merge}
\alias{exp.merge}
\title{Merge from exp}
\usage{
\method{exp}{merge}(exp, expect_repeats = FALSE, merge_passage = FALSE)
}
\arguments{
\item{exp}{char: like "240/480/240"}

\item{expect_repeats}{bool: whether repeat titers are expected}

\item{merge_passage}{bool: not implemented}
}
\value{
char: merged titer table
}
\description{
Get merged titer table from single experiment.
}
