% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_titertables.R
\name{expdb.longTibble}
\alias{expdb.longTibble}
\title{Parse an expdb to tibble}
\usage{
expdb.longTibble(expdb)
}
\arguments{
\item{expdb}{}
}
\value{
tibble
}
\description{
Parse an experiment record into a long tibble containing all results.
}
