% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_merging.R
\name{expdb.merge}
\alias{expdb.merge}
\title{Merge from expdb}
\usage{
expdb.merge(expdb, duplicate_titers = "mean")
}
\arguments{
\item{expdb}{char: like "240/480/240"}

\item{duplicate_titers}{char: "mean" or "combine"}
}
\value{
char: merged titer table
}
\description{
Get merged titer table from experimental database.
}
