% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_parsing.R
\name{exper.longTibble}
\alias{exper.longTibble}
\title{Parse an experiment to tibble}
\usage{
exper.longTibble(exp)
}
\arguments{
\item{exp}{}
}
\value{
tibble
}
\description{
Parse an experiment record into a long tibble containing all results.
}
