% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_merging.R
\name{exper.merge}
\alias{exper.merge}
\title{Merge from exp}
\usage{
exper.merge(
  exp,
  expect_repeats = FALSE,
  duplicate_titers = "mean",
  merge_passage = FALSE,
  lowerthreshold = 10,
  upperthreshold = Inf,
  threshold_masking = F
)
}
\arguments{
\item{exp}{char: like "240/480/240"}

\item{expect_repeats}{bool: whether repeat titers are expected}

\item{merge_passage}{bool: not implemented}

\item{threshold}{}
}
\value{
array of char: merged titer table
}
\description{
Get merged titer table from single experiment.
}
