% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_titertables.R
\name{exper.toLongTibble}
\alias{exper.toLongTibble}
\title{Parse an experiment to tibble}
\usage{
exper.toLongTibble(exp)
}
\arguments{
\item{exp}{}
}
\value{
tibble
}
\description{
Parse an experiment record into a long tibble containing all results.
}
