% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_methods.R
\name{\%$\%}
\alias{\%$\%}
\title{Subsetting for databases}
\usage{
db \%$\% val
}
\description{
Returns field for all database entries. Multilayer subsetting is also supported with \code{db\%$\%field$subfield}
}
