% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{h3_clade_grouping_heirarchy}
\alias{h3_clade_grouping_heirarchy}
\title{order for matching}
\format{
An object of class \code{character} of length 7.
}
\usage{
h3_clade_grouping_heirarchy
}
\description{
order for matching
}
\keyword{datasets}
