% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{h3_clade_order}
\alias{h3_clade_order}
\title{order for plotting}
\format{
An object of class \code{character} of length 9.
}
\usage{
h3_clade_order
}
\description{
order for plotting
}
\keyword{datasets}
