% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{is.substitution}
\alias{is.substitution}
\title{Check a substitutions entry}
\usage{
is.substitution(subs)
}
\arguments{
\item{subs}{char: A proposed substitution entry}
}
\value{
character
}
\description{
Checks if a proposed substitution follows the correct formatting
}
\details{
The following rules are currently checked:
\enumerate{
\item substitution is of character type
\item starts and ends with valid single letter amino acid code
\item central characters form a numerb in 1:484
}
}
