% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_merging.R
\name{logtiter.toPlot}
\alias{logtiter.toPlot}
\title{this converts logtiters with less thans to logtiters where <3 -> 2}
\usage{
logtiter.toPlot(titers)
}
\description{
this converts logtiters with less thans to logtiters where <3 -> 2
}
