% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_titertables.R
\name{longtiters.addRecords}
\alias{longtiters.addRecords}
\title{Add ag, sr, exp records from IDs to tibble}
\usage{
longtiters.addRecords(
  longtiters,
  agdb = get_agdb(),
  srdb = get_srdb(),
  expdb = get_expdb()
)
}
\arguments{
\item{longtiters}{}
}
\value{
tibble
}
\description{
Add ag, sr, exp records from IDs to tibble
}
