% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_titertables.R
\name{longtiters.plotdata}
\alias{longtiters.plotdata}
\title{Add tibble plotting columns}
\usage{
longtiters.plotdata(titertable_long, agdb = acutilsLite:::get_agdb())
}
\arguments{
\item{titertable_long}{tibble}
}
\value{
tibble
}
\description{
Add columns useful for plotting titer data
}
