% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdatabase_functions.R
\name{match_hi_name}
\alias{match_hi_name}
\title{functions for working wiht eugene's sequence databases

Return the seuqnece entry corresponding to a given hi name (v$s[\link{idx}]$h)}
\usage{
match_hi_name(name, ...)
}
\description{
functions for working wiht eugene's sequence databases

Return the seuqnece entry corresponding to a given hi name (v$s[\link{idx}]$h)
}
