% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdatabase_functions.R
\name{match_ref}
\alias{match_ref}
\title{Find sequence entry with matching R (hash and name)}
\usage{
match_ref(R, ..., ignore_name = F, return_full_virus = F)
}
\description{
Find sequence entry with matching R (hash and name)
}
