% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_merging.R
\name{mergeTiters}
\alias{mergeTiters}
\title{Merge titers}
\usage{
mergeTiters(
  titers,
  lowerthreshold = 10,
  upperthreshold = Inf,
  threshold_masking = F
)
}
\arguments{
\item{titers}{char: like "240/480/240"}
}
\value{
char: merged titer
}
\description{
Merge titers (formatted like "240/480/240") by taking mean of logs (and exponentiating back)
}
\details{
When logs are taken <40 is converted to 20 for averaging
}
