% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{outbox.agdb.database}
\alias{outbox.agdb.database}
\title{Mark which properties should be unboxed}
\usage{
outbox.agdb.database(db.database)
}
\arguments{
\item{db.database}{list}
}
\value{
list
}
\description{
Helper for write.agdb
}
