% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_sera.R
\name{outbox.srdb}
\alias{outbox.srdb}
\title{Mark which properties should be unboxed}
\usage{
outbox.srdb(db)
}
\arguments{
\item{db.database}{list}
}
\value{
list
}
\description{
Helper for write.srdb
}
