% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{place_abvs}
\alias{place_abvs}
\title{All known two-letter place name abbreviations}
\format{
An object of class \code{character} of length 629.
}
\usage{
place_abvs
}
\description{
All known two-letter place name abbreviations
}
\keyword{datasets}
