% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_methods.R
\name{print.acdatabase.entry}
\alias{print.acdatabase.entry}
\title{Printing an acdb entry}
\usage{
\method{print}{acdatabase.entry}(db.ag)
}
\arguments{
\item{db.ag}{list}
}
\value{
character
}
\description{
Formatted print for acdatabase entries
}
