% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agdb.R
\name{read.agdb}
\alias{read.agdb}
\title{Read antigen database}
\usage{
read.agdb(file)
}
\arguments{
\item{file}{character}
}
\value{
??? agdb object
}
\description{
Read antigen database
}
\examples{
read.agdb('../path/to/db.json')
}
