% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabases_plot.R
\name{sequences_plot}
\alias{sequences_plot}
\title{Make sequence comparison plot}
\usage{
sequences_plot(
  sequences,
  mask = rep(1, length(sequences)),
  mutagenised_pos = NULL
)
}
\description{
sequences_tibble should have:
\enumerate{
\item a row per antigen, with its name and id
\item its sequence
\item the id of its parent
}
}
