% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{sr.attribute}
\alias{sr.attribute}
\title{Get serum attribute}
\usage{
sr.attribute(sr, attribute, inherit = T)
}
\arguments{
\item{sr}{environment: a serum}

\item{attribute}{char: the name of the desired attiribute}

\item{inherit:}{bool: whether to inherit attributes from parents}
}
\value{
attribute value
}
\description{
Find an attribute value for an serum By default this is set to check parent sera if attribute not found at lower level.
}
