% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{sr.cluster}
\alias{sr.cluster}
\title{Get serum cluster}
\usage{
sr.cluster(sr, agdb)
}
\arguments{
\item{identifier}{char: an id, long, etc}

\item{from:}{char: which attribute to search with (long, id, etc)}
}
\value{
char
}
\description{
Gets the cluster of a serum's homologous antigen
}
