% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{sr.experiments}
\alias{sr.experiments}
\title{Get experiments that a serum is in}
\usage{
sr.experiments(sr, expdb, include_passage_variants = FALSE, srdb)
}
\description{
Get experiments that a serum is in
}
