% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{sr.inheritance}
\alias{sr.inheritance}
\title{Get inheritance list for a serum.}
\usage{
sr.inheritance(sr)
}
\arguments{
\item{sr}{environment: a serum}
}
\value{
list
}
\description{
Get a list of parent sera (previous sera from same individual).
}
