% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_sera.R
\name{srdb.append}
\alias{srdb.append}
\title{Append a serum to an srdb}
\usage{
srdb.append(srdb, sr)
}
\arguments{
\item{srdb}{list}

\item{sr}{environment}
}
\value{
list
}
\description{
Append a serum to an srdb
}
