% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{srdb.clusters}
\alias{srdb.clusters}
\title{Get sera clusters}
\usage{
srdb.clusters(sera, agdb, srdb)
}
\arguments{
\item{sera}{vector of identifiers}

\item{from:}{char: which attribute to search with (long, id, etc)}
}
\value{
char
}
\description{
Gets the clusters of sera's homologous antigens.
}
