% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{srdb.getIDs}
\alias{srdb.getIDs}
\title{Subset serum database based on IDs}
\usage{
srdb.getIDs(ids, srdb)
}
\arguments{
\item{ids}{}
}
\value{
list
}
\description{
Throws error if some IDs not found
}
