% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{srdb.homologousAntigens}
\alias{srdb.homologousAntigens}
\title{Get homologous sera}
\usage{
srdb.homologousAntigens(srdb, agdb)
}
\arguments{
\item{agdb}{}
}
\value{
list
}
\description{
Gets homologous antigens for all sera in an srdb by matching ids.
}
