% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_sera.R
\name{srdb.id}
\alias{srdb.id}
\title{Generate an id}
\usage{
srdb.id(db, id = NULL, excluded_ids = NULL)
}
\arguments{
\item{id}{should be left NULL}

\item{excluded_ids}{char (optional): a vector of ids which should not be assigned}

\item{srdb}{list a serum databses (?srdb)}
}
\value{
character
}
\description{
Generates a random id (6 capital letters and numbers) which is not seen in the provided srdb or excluded_ids
}
