% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{srdb.matchAnimalIDsIndices}
\alias{srdb.matchAnimalIDsIndices}
\title{Find sera matching an animal id}
\usage{
srdb.matchAnimalIDsIndices(srdb, animal_id)
}
\description{
Find sera matching an animal id
}
