% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_sera.R
\name{srdb.merge}
\alias{srdb.merge}
\title{Merge sera in serum database}
\usage{
srdb.merge(srdb, sr_id_keep, sr_id_remove)
}
\arguments{
\item{sr_id_keep}{char}

\item{sr_id_remove}{char}

\item{agdb}{list}
}
\value{
list
}
\description{
Merge sera in serum database
}
