% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_sera.R
\name{srdb.new}
\alias{srdb.new}
\title{Create serum database}
\usage{
srdb.new(db = list())
}
\arguments{
\item{db}{list}
}
\value{
environment
}
\description{
Creates a serum database from a list of antigen entries. For information on database structure, see ?srdb.
}
