% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_sera.R
\name{srdb}
\alias{srdb}
\title{Serum database}
\description{
Serum database
}
\details{
The serum database is ultimately stored as a json file with fields:

id
long
strain_id
parent_id
species
animal_id
sample_date
meta
comments
aliases

Not all fields are filled for all sera

In R, this is converted to a list of environments (see \code{?sr}) with the above fields, plus:
\enumerate{
\item A ar$.parent entry, which is the database entry for the parent serum
}

Supports subsetting with \verb{\%$\%} (see \verb{?\%$\%})
}
