% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_sera.R
\name{srdb.sr}
\alias{srdb.sr}
\title{Produce a serum entry for the database}
\usage{
srdb.sr(
  id,
  long,
  parent_id,
  strain_id,
  species,
  animal_id,
  sample_date,
  meta,
  aliases,
  srdb,
  .parent = NULL
)
}
\value{
character
}
\description{
Produces an entry for serum databases (see ?srdb) from the passed data.
}
