% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{srid.cluster}
\alias{srid.cluster}
\title{Get serum cluster}
\usage{
srid.cluster(identifier, agdb, srdb, from = "id")
}
\arguments{
\item{identifier}{char: an id, long, etc}

\item{from:}{char: which attribute to search with (long, id, etc)}
}
\value{
char
}
\description{
Gets the cluster of a serum's homologous antigen
}
