% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_names.R
\name{standardizeStrainName}
\alias{standardizeStrainName}
\title{Function for standardising names locally}
\usage{
standardizeStrainName(
  name,
  default_species = NA,
  default_virus_type = "A",
  default_virus_subtype = "HXNX",
  stop_on_error = TRUE
)
}
\description{
Function for standardising names locally
}
