% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_titertables.R
\name{summarise_plotdata_by_ag_cluster}
\alias{summarise_plotdata_by_ag_cluster}
\title{Summarise by antigen cluster}
\usage{
summarise_plotdata_by_ag_cluster(plotdata)
}
\arguments{
\item{plotdata}{tibble}
}
\value{
tibble
}
\description{
Summarise by antigen cluster
}
