% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_parsing.R
\name{summarise_plotdata_by_sr_cluster}
\alias{summarise_plotdata_by_sr_cluster}
\title{Summarise by serum cluster}
\usage{
summarise_plotdata_by_sr_cluster(plotdata)
}
\arguments{
\item{plotdata}{tibble}
}
\value{
tibble
}
\description{
Summarise by serum cluster
}
