% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_titertables.R
\name{tibble.factorize}
\alias{tibble.factorize}
\title{convert column(s) into factor for plotting}
\usage{
tibble.factorize(tib, columns)
}
\description{
convert column(s) into factor for plotting
}
