% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_titertables.R
\name{titerlong.plotdata}
\alias{titerlong.plotdata}
\title{Add tibble plotting columns}
\usage{
titerlong.plotdata(titertable_long)
}
\arguments{
\item{titertable_long}{tibble}
}
\value{
tibble
}
\description{
Add columns useful for plotting titer data
}
