% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_antigens.R
\name{write.agdb}
\alias{write.agdb}
\title{Write database to file}
\usage{
write.agdb(db, file)
}
\arguments{
\item{file}{char}

\item{agdb}{list}
}
\value{
list
}
\description{
Writes an agdb to specified location. If a database with the same name already exists at that location, provides summary of additions, deletions and edits to database
}
