% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_sera.R
\name{write.srdb}
\alias{write.srdb}
\title{Write database to file}
\usage{
write.srdb(db, file)
}
\arguments{
\item{file}{char}

\item{agdb}{list}
}
\value{
list
}
\description{
Writes an agdb to specified location.
}
