% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{acdb.Long}
\alias{acdb.Long}
\title{Select long name from antigen entries}
\usage{
acdb.Long(acdb)
}
\arguments{
\item{ags}{}
}
\value{
Returns a vector of long names
}
\description{
Select long name from antigen entries
}
