% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.clade}
\alias{ag.clade}
\title{Get antigen clade
You can choose whether clade is searched for only in ag (how = 'self), only in the root virus (how = 'root), or anywhere in the inheritance (how = 'any')}
\usage{
ag.clade(ag, how = "self", agdb = acutilsLite:::get_agdb())
}
\arguments{
\item{how}{= c('self', 'root', 'any')}
}
\value{
char
}
\description{
Get antigen clade
You can choose whether clade is searched for only in ag (how = 'self), only in the root virus (how = 'root), or anywhere in the inheritance (how = 'any')
}
