% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_utils.R
\name{ag.sequence}
\alias{ag.sequence}
\title{Get sequence for gene}
\usage{
ag.sequence(ag, gene = "HA", agdb = acutils:::get_agdb(), inherit = T)
}
\description{
If sequence not present in ag entry, looks in parent antigen entries for sequence, and applies substitutions
}
