% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdatabase_merging.R
\name{mergeTiters}
\alias{mergeTiters}
\title{Handling titers:}
\usage{
mergeTiters(titers, threshold = 10)
}
\arguments{
\item{titers}{char: like "240/480/240"}
}
\value{
char: merged titer
}
\description{
Handling titers:
}
\details{
Merge titers

Merge titers (formatted like "240/480/240") by taking mean of logs (and exponentiating back)

When logs are taken <40 is converted to 20 for averaging
}
